\name{calWeight}
\alias{calWeight}
\title{calWeight}
\usage{
  calWeight(aode)
}
\arguments{
  \item{aode}{list. this is the list which has all the
  required variables in it.}
}
\value{
  aode list. updated value
}
\description{
  This function calculates mutual information between the
  attr and class.
}
\details{
  This function is called when weighted flag is set
}
\author{
  saiteja ranuva
}

