\name{distributionForInstance}
\alias{distributionForInstance}
\title{distributionForInstance}
\usage{
  distributionForInstance(x, aode)
}
\arguments{
  \item{x}{instance to be classified}

  \item{aode}{list. this is the list which has all the
  required variables in it.}
}
\value{
  class integer predicted class of the instance
}
\description{
  predicts class of a given instance based on the model
}
\details{
  details to be added
}
\author{
  sai teja ranuva
}

