\name{indexCalc}
\alias{indexCalc}
\title{fun1}
\usage{
  indexCalc(aode)
}
\arguments{
  \item{aode}{list. this is the list which has all the
  required variables in it.}
}
\value{
  a list
}
\description{
  Calculates all the array and matrix indices required.
}
\details{
  Details - add later
}
\author{
  sai teja ranuva
}
\seealso{
  \code{\link{setVar}}
}

