\name{setVar}
\alias{setVar}
\title{setVar}
\usage{
  setVar(aode)
}
\arguments{
  \item{aode}{list. this is the list which has all the
  required variables in it.}
}
\value{
  aode list updated value of the list
}
\description{
  sets the required variables
}
\details{
  calculates the required space for indices and allocates
}
\author{
  sai teja ranuva
}

