\name{training}
\alias{training}
\title{training}
\usage{
  training(aode)
}
\arguments{
  \item{aode}{list. this is the list which has all the
  required variables in it.}
}
\value{
  aode list. updated value
}
\description{
  Calculates the count matrices.
}
\details{
  class parent count and class parent child count are
  calculated
}
\author{
  sai teja ranuva
}

